chore: column_field_helper file rename
